Move transmutes_expressible_as_ptr_casts to its own module
